+++ packages/protocolbeat/src/apps/discovery/panel-nodes/store/actions/group.ts
@@ -1,6 +1,7 @@
 import { unique } from '@l2beat/shared-pure'
 import type { Field, Node, State } from '../State'
 import { NODE_WIDTH } from '../utils/constants'
+import type { StoredGroup } from '../utils/layout'
 import { updateNodePositions } from '../utils/updateNodePositions'
 
 export function groupSelected(state: State): Partial<State> {
@@ -53,30 +54,49 @@ export function renameSelectedGroup(
   return { nodes }
 }
 
-function createPhantomNode(subnodes: Node[]): Node {
-  const anchor = subnodes[0]?.box
-  const internal = unique(subnodes.flatMap((n) => collectIds(n)))
+function createPhantomNode(members: Node[]): Node {
+  const anchor = members[0]?.box
+  return makeGroupNode(
+    {
+      id: `group:${crypto.randomUUID()}`,
+      name: 'Group',
+      opened: false,
+      color: 0,
+      box: {
+        x: anchor?.x ?? 0,
+        y: anchor?.y ?? 0,
+        width: NODE_WIDTH,
+        height: NODE_WIDTH,
+      },
+      members: [],
+    },
+    members,
+  )
+}
+
+export function makeGroupNode(group: StoredGroup, members: Node[]): Node {
+  const internal = unique(members.flatMap((n) => collectIds(n)))
   return {
-    id: `group:${crypto.randomUUID()}`,
+    id: group.id,
     address: '',
     isInitial: false,
     hasTemplate: false,
     addressType: 'Group',
-    name: 'Group',
-    fields: collectOutgoingFields(subnodes, internal),
+    name: group.name,
+    fields: collectOutgoingFields(members, internal),
     hiddenFields: [],
     box: {
-      x: anchor?.x ?? 0,
-      y: anchor?.y ?? 0,
-      width: NODE_WIDTH,
-      height: NODE_WIDTH,
+      x: group.box.x,
+      y: group.box.y,
+      width: group.box.width ?? NODE_WIDTH,
+      height: group.box.height ?? NODE_WIDTH,
     },
-    color: 0,
+    color: group.color,
     hueShift: 0,
     data: null,
     isReachable: true,
-    opened: false,
-    subnodes,
+    opened: group.opened,
+    subnodes: members,
   }
 }
